Add the chosen quality setting to the output filename.
Previously, the output name was the same as used for the temporary
encode directory. However, a given encode directory may contain several
different quality levels as part of the target quality feature, so the
output filename needs to add the chosen quality to distinguish it from
other potential encodes that otherwise use the same encoding settings.

Signed-off-by: Jason Lynch <jason@aexoden.com>