Allow target quality metric calculations to take into account the number of active workers.
In a case where the scene lengths are very disparate, it's possible to
have one or two workers that end up working long beyond everything else
is already done. To counter this somewhat, this allows them to use more
threads for the metric calculations. This especially benefits
SSIMULACRA2 which isn't particularly fast to begin with.

Signed-off-by: Jason Lynch <jason@aexoden.com>