Increase the minimum CRF used for x264.
This is primarily to prevent x264 from entering true lossless mode for
the entire clip, which causes mkvmerge to complain about appending it to
non-lossless clips. Practically speaking, nobody would want files with
CRF values this low anyway (as the bitrate is astronomically high)
unless they were intending on doing a truly lossless encode in the first
place. And in that case, they wouldn't use this tool.

Signed-off-by: Jason Lynch <jason@aexoden.com>